Initial commit of gexport plugin.  Note that it's not entirely functional yet.
This plugin will replace the native graph export in Cacti.  It is
designed to allow you to have multiple export definitions.  In phase II
of this plugin development, additional features will be included to
provide access control to authorized users.  There are additional
features in this version of the plugin, but again, it's not finished
yet.